backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
email-validator>=2.2.0
